Feat/condvar

+++ src/ffi/thread.rs
@@ -1,7 +1,7 @@
 use crate::ThreadId;
 use crate::core::detector::thread::{on_thread_exit, on_thread_spawn};
 use crate::core::get_current_thread_id;
-use std::os::raw::{c_int, c_ulong};
+use std::os::raw::c_int;
 
 /// Register a thread spawn with the global detector.
 ///
@@ -20,8 +20,8 @@ use std::os::raw::{c_int, c_ulong};
 /// - This function is normally called automatically by the CREATE_TRACKED_THREAD macro.
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn deloxide_register_thread_spawn(
-    thread_id: c_ulong,
-    parent_id: c_ulong,
+    thread_id: usize,
+    parent_id: usize,
 ) -> c_int {
     let parent = if parent_id == 0 {
         None
@@ -47,7 +47,7 @@ pub unsafe extern "C" fn deloxide_register_thread_spawn(
 /// - The caller must ensure thread_id represents a real thread that is exiting.
 /// - This function is normally called automatically by the CREATE_TRACKED_THREAD macro.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn deloxide_register_thread_exit(thread_id: c_ulong) -> c_int {
+pub unsafe extern "C" fn deloxide_register_thread_exit(thread_id: usize) -> c_int {
     on_thread_exit(thread_id as ThreadId);
     0 // Success
 }
@@ -63,6 +63,6 @@ pub unsafe extern "C" fn deloxide_register_thread_exit(thread_id: c_ulong) -> c_
 /// # Safety
 /// This function is safe to call from FFI contexts.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn deloxide_get_thread_id() -> c_ulong {
-    get_current_thread_id() as c_ulong
+pub unsafe extern "C" fn deloxide_get_thread_id() -> usize {
+    get_current_thread_id() as usize
 }
